backend/requirements.txt
# FinAgent Pro - Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# AI/ML Libraries
openai==1.3.5
anthropic==0.7.1
scikit-learn==1.3.2
prophet==1.1.5
xgboost==2.0.2
tensorflow==2.15.0

# OCR & Document Processing
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
google-cloud-vision==3.4.5

# Data Processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Fraud Detection
isolation-forest==0.1.1
hdbscan==0.8.33

# Payment Integration
stripe==7.6.0

# PDF Generation
reportlab==4.0.7
weasyprint==60.1

# Communication
sendgrid==6.11.0
twilio==8.10.3

# Monitoring & Logging
prometheus-client==0.19.0
loguru==0.7.2
sentry-sdk==1.38.0

# Task Queue
celery==5.3.4
kombu==5.3.4

# API Documentation
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.25.2
aiofiles==23.2.1
